Add files via upload
I was able to fit a total of 7 visuals in the dashboard. This being said, more can fit if we decide to remove one of the slicers as well as the score cards. I can also fit one more regardless. For the interactive features, we currently have 3: drill down on the ethnicity and gender visual, 2 slicers with buttons/tiles, and data highlighting actions. Also, we used 2 kinds of filters: slicers and the side pane filters (not sure if this counts). In addition, we now have a total of 14 visuals as I created a variation of each visual on the dashboard (expect for the response time which Sabrina had created).